Renamed formatting module to utils
