+++ pynote/utils.py
